requirements.txt
Django~=3.2.9
Pillow~=8.4.0

djangorestframework~=3.12.4